@@ -1,8 +1,9 @@
 {
+  "type": "module",
+  "main": "./dist/ts2fable.js",
   "files": [
     "dist/"
   ],
-  "main": "./dist/ts2fable.js",
   "bin": {
     "ts2fable": "./dist/ts2fable.js"
   },
@@ -10,8 +11,8 @@
   "version": "0.0.0",
   "description": "TypeScript definition files parser for fable-compiler",
   "scripts": {
-    "ts2fable": "node --require esm ./build/cli/ts2fable.js",
-    "test": "mocha --require esm ./build/test/test.js"
+    "ts2fable": "node ./build/cli/ts2fable.js",
+    "test": "mocha ./build/test/test.js"
   },
   "repository": {
     "type": "git",
@@ -35,66 +36,44 @@
   },
   "devDependencies": {
     "@babel/core": "^7.13.10",
-    "@babel/plugin-external-helpers": "^7.12.13",
     "@babel/plugin-proposal-class-properties": "^7.13.0",
-    "@babel/plugin-transform-modules-commonjs": "^7.13.8",
     "@babel/plugin-transform-runtime": "^7.13.10",
     "@babel/polyfill": "^7.12.1",
     "@babel/preset-env": "^7.13.10",
     "@babel/preset-react": "^7.12.13",
-    "@babel/register": "^7.13.8",
-    "@types/chai": "^4.2.14",
-    "@types/google-protobuf": "^3.7.4",
-    "@types/jquery": "^3.5.5",
-    "@types/mocha": "^2.2.44",
-    "@types/node": "^8.10.66",
-    "@types/react": "16.0.4",
-    "@types/react-native": "^0.56.4",
-    "@types/yargs": "^13.0.11",
+    "@monaco-editor/react": "^4.3.1",
+    "@types/mocha": "^9.0.0",
+    "@types/node": "^16.11.11",
+    "@types/react": "^17.0.37",
     "babel-loader": "^8.2.2",
-    "breeze-client": "^1.7.1",
-    "bulma": "^0.7.4",
+    "bulma": "^0.9.3",
     "bulma-pageloader": "^2.1.0",
     "bulma-tooltip": "^2.0.1",
-    "chai": "^4.1.2",
-    "chalk": "^2.3.0",
     "clipboard": "^2.0.6",
-    "copy-webpack-plugin": "^8.0.0",
-    "cross-env": "^5.1.1",
-    "css-loader": "^5.1.3",
-    "electron": "^11.5.0",
-    "eslint-scope": "3.7.1",
-    "esm": "^3.2.25",
+    "copy-webpack-plugin": "^10.0.0",
+    "css-loader": "^6.5.1",
     "file-loader": "^6.2.0",
-    "font-awesome": "^4.7.0",
-    "gh-pages": "^1.2.0",
     "html-webpack-plugin": "^5.3.1",
-    "izitoast": "^1.1.5",
-    "mini-css-extract-plugin": "^1.3.9",
-    "mocha": "^6.2.3",
-    "monaco-editor": "^0.21.3",
-    "node-sass": "^5.0.0",
-    "office-ui-fabric-react": "^5.135.6",
+    "mini-css-extract-plugin": "^2.4.4",
+    "mocha": "^9.1.3",
+    "monaco-editor": "^0.31.0",
     "path-browserify": "^1.0.1",
     "prepend-file": "^2.0.0",
     "process": "^0.11.10",
     "prop-types": "^15.6.2",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "react-monaco-editor": "0.13.0",
-    "react-resize-detector": "^3.0.1",
-    "reactxp": "^1.0.2",
-    "requirejs": "^2.3.5",
-    "rollup": "^2.42.1",
-    "sass-loader": "^11.0.1",
-    "source-map-loader": "^2.0.1",
-    "style-loader": "^2.0.0",
-    "vscode": "^1.1.37",
-    "webpack": "^5.27.1",
-    "webpack-cli": "^4.5.0",
-    "webpack-dev-server": "^3.11.2"
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-resize-detector": "^6.7.6",
+    "rollup": "^2.60.2",
+    "sass": "^1.43.4",
+    "sass-loader": "^12.3.0",
+    "source-map-loader": "^3.0.0",
+    "style-loader": "^3.3.1",
+    "webpack": "^5.64.4",
+    "webpack-cli": "^4.9.1",
+    "webpack-dev-server": "^4.6.0"
   },
   "engines": {
-    "vscode": "^1.18.1"
+    "node": "^12.20.0 || ^14.13.1 || >=16.0.0"
   }
 }
