@@ -1,3 +1,3 @@
 // https://www.npmjs.com/package/prepend-file
-var prependFile = require('prepend-file');
+import prependFile from "prepend-file"
 prependFile('./dist/ts2fable.js', '#!/usr/bin/env node\n')
