@@ -4,12 +4,14 @@ open System
 open Fable.Core
 open Fable.Core.JS
 
+#nowarn "3390"
+
 type Array<'T> = System.Collections.Generic.IList<'T>
 type ReadonlyArray<'T> = System.Collections.Generic.IReadOnlyList<'T>
 type ReadonlyMap<'K, 'V> = Map<'K, 'V>
 type Symbol = obj
 
-let [<Import("*","typescript")>] ts: Ts.IExports = jsNative
+let [<ImportDefault("typescript")>] ts: Ts.IExports = jsNative
 
 type [<AllowNullLiteral>] IExports =
     abstract setTimeout: handler: (ResizeArray<obj option> -> unit) * timeout: float -> obj option
